requirements.txt
certifi==2019.11.28
chardet==3.0.4
gunicorn==20.0.4
idna==2.8
paho-mqtt==1.5.0
pyTelegramBotAPI==3.6.6
requests==2.10.0
six==1.13.0
threaded==4.0.8
urllib3==1.25.7
